Major overhaul
Discovered unusual behavior due to the use of sprintf, couldn't find a way to make it work properly so I ditched that in place of my own code that does essentially the same thing. In the process, I found that in some of the code where there is an XOR or Addition % 128, the resulting string char can end up becoming a null byte, which terminates the string prematurely when using functions like strlen. After much debugging and testing, it seems that it all works on both Ubuntu and Kali, but I'm not 100% confident that there is no other kind of bugs lurking in there.